+++ contracts/nexus/Nexus.sol
@@ -81,7 +81,7 @@ contract Nexus is INexus, ModuleKeys, DelayedClaimableGovernor {
         require(len == _addresses.length, "Insuffecient address data");
         require(len == _isLocked.length, "Insuffecient locked statuses");
 
-        for(uint i = 0 ; i < len; i++) {
+        for(uint256 i = 0 ; i < len; i++) {
             _publishModule(_keys[i], _addresses[i], _isLocked[i]);
         }
 
@@ -158,7 +158,7 @@ contract Nexus is INexus, ModuleKeys, DelayedClaimableGovernor {
         uint256 len = _keys.length;
         require(len > 0, "Keys array empty");
 
-        for(uint i = 0 ; i < len; i++) {
+        for(uint256 i = 0 ; i < len; i++) {
             _acceptProposedModule(_keys[i]);
         }
     }
